Add files via upload
Made some changes to song_scraper.py to replace unicode characters with spaces.  Otherwise there were errors and prevention of songs with apostrophes or non-English letters to be added. Also included url for Just Dance 2020 and third page of unlimited  songs.

Made change to song_chooser.py to ask user for number of songs so that a playlist could be made.